A Xi Jinping Quote Scraper
This is a simple webscraper which extracts the daily Xi Jinping quote published on 'http://www.chinadaily.com.cn/china/xismoments'. It takes the date published and the text, and adds them to a .txt file. When enough quotes have been analysed, further analysis of these quotes can be undertaken.